If sfi is undefined = '0' not 0
An error was picked up when line 275 tried to
trim() the var sfi. However, it must have been an
empty field or the whole XML did not come back,
because sfi was set to 0 based on the IF/THENs
that grabbed the XML and stored it as variables or
created a default value if it was undefined.
However a default of 0 cannot be trimmed, so it
needed to be changed to '0'. While I was at it,
I created a default value if timestamp ended up
being undefined. Also put in a simplification
of the IF/THEN statements. Both the bug fix and the
simplification are thanks to echicken.